requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

numpy==2.4.2
pandas==3.0.1
Pillow==12.1.1
prettytable==3.17.0
pytorch-lightning==2.6.1
pytorch-metric-learning==2.9.0
scikit-learn==1.8.0
timm==1.0.25
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.3
transformers==4.57.3
# Optional: DINOv2 uses xformers when available, but the code falls back cleanly without it.
faiss-gpu-cu12==1.14.1.post1
